seed-dummy-data/pyproject.toml
[project]
name = "seed-dummy-data"
version = "0.1.0"
description = "Seed comprehensive demo data for the Value platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "clickhouse-connect>=0.7.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "uuid7>=0.1.0",
    "psycopg2-binary>=2.9.0",
    "value-python>=0.1.1",
]

[project.scripts]
seed-data = "seed_data:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = ["ruff>=0.1.0", "mypy>=1.0.0"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
